+++ src/ManifoldMeshes.jl
@@ -32,7 +32,7 @@ export all_cell_volumes, all_node_coordinates, all_cell_centroids, all_edge_leng
 export boundary_nodes, boundary_edges
 export dual, has_dual, cell_face, cell_local_2d
 export slerp, node_points, edge_segments, cell_polygons, cell_triangles
-export spherical_triangle_area, spherical_polygon_area
+export spherical_triangle_area, spherical_polygon_area, spherical_polygon_intersection
 export side_of_geodesic, geodesic_arc_intersection, cell_ring
 
 include("sphere/utils.jl")